Organize agent participant backend functions
